✨ refactor(lang/init): Correct Ender Ore naming and update imports
Updates the lang file and blockstate JSON for Ender Ore to use the correct name 'ender_ore'.
Also, updates imports in OKProgressionsItems and OKProgressionsBlocks to reference the correct 'Reference' class.
