+++ src/main/resources/assets/okprogressions/blockstates/ender_ore.json
